Conflicts:
	build/build/conf.js
	build/build/pack.js

+++ build/build/pack.js
@@ -101,12 +101,20 @@ module.exports = function (src, baton) {
     copyFolder(_c.CLIENTFILES, clientFilesDest);
     copyFolder(_c.DEPENDENCIES_BOOTSTRAP, bootstrapDest);
     copyFolder(_c.DEPENDENCIES_JNEXT, jnextDest);
-    copyFolder(_c.DEPENDENCIES_WEBPLATFORM_UI, _c.DEPLOY_UI);
+    
 
-    //Copy files to target directory (DO NOT copy webplatform-framework lib/* files over)
-    utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_FRAMEWORK_REQUIRE, browserRequireDest);
-    utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_FRAMEWORK_LIB, webplatformDest);
-    utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_I18N, webplatformDest);
+    //Copy webplatform files/folders if they exist
+    if (path.existsSync(_c.DEPENDENCIES_WEBPLATFORM)) {
+		copyFolder(_c.DEPENDENCIES_WEBPLATFORM_UI, _c.DEPLOY_UI);
+        utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_FRAMEWORK_REQUIRE, browserRequireDest);
+        utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_FRAMEWORK_LIB, webplatformDest);
+		utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_I18N, webplatformDest);
+    } else {
+        console.log("\n****ERROR: Webplatform could not be found.****");
+        console.log("Please refer to the README for instructions on how to copy the Webplatform from a BB10 Webworks SDK installation.\n");
+    }
+	
+	//Copy files to target directory (DO NOT copy webplatform-framework lib/* files over)
     utils.copyFile(readmeFile, _c.DEPLOY);
     utils.copyFile(licenseFile, _c.DEPLOY);
